% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mcadet_function.R
\name{mcadet}
\alias{mcadet}
\title{Mcadet: Feature Selection for scRNA-seq Data}
\usage{
mcadet(
  data,
  n.comp = 60,
  run = 10,
  n.feature = NA,
  nk_percent = 0.005,
  start_resolution = 0.5,
  cell_percent = 0.005,
  MC_iter = 50000,
  fdr = 0.15,
  seed = 1234
)
}
\arguments{
\item{data}{A raw count matrix (or data frame) where rows represent genes and columns represent cells. (must have row and column names)}

\item{n.comp}{The number of principal components (PCs) selected from MCA decomposition. Default is 60.}

\item{run}{The number of iterations to run for a more robust result. Default is 10.}

\item{n.feature}{The number of genes to select. Default is NA, which means the proposed statistical testing method will be used.}

\item{start_resolution}{The starting resolution parameter for the Leiden algorithm, increasing by 0.1 for each run. Default is 0.5.}

\item{cell_percent}{The percentage threshold for filtering low-expressed genes. Genes expressed in fewer cells than this threshold will be excluded. Default is 0.005.}

\item{MC_iter}{Monte-Carlo iteration times. Default is 50,000.}

\item{fdr}{The false discovery rate (FDR) threshold for the statistical testing. Default is 0.15.}

\item{seed}{The random seed for the Leiden algorithm. Default is 1234.}
}
\value{
The function returns a `data.frame` with three columns:
\itemize{
  \item \strong{gene}: The name of each selected gene.
  \item \strong{logR}: The log ratio (maximum/minimum rank) of each gene's ranking across iterations.
  \item \strong{p.value}: The p-value obtained from Monte-Carlo simulations, representing the significance of each gene.
}
}
\description{
This function performs feature selection for scRNA-seq (count-based) data. It integrates Multiple Correspondence Analysis (MCA), graph-based community detection, and a novel statistical testing approach to select the most significant features (genes).
}
\examples{
\dontrun{
data <- matrix(rnbinom(10000000,2,0.5), nrow = 10000, ncol = 1000)
rownames(data) = paste0("Gene_", c(1:10000))
colnames(data) = paste0("Cell_", c(1:1000))
selected_genes <- mcadet(data)
}

}
